Moved flightInfo logic into AirService
