+++ src/Services/Air/AirService.js
@@ -69,5 +69,17 @@ module.exports = (settings) => {
           });
       });
     },
+
+    flightInfo(options) {
+      const AirService = airServiceInternal(auth, debug, production);
+      return AirService.flightInfo(options)
+        .then(data => data)
+        .catch((err) => {
+          if (debug > 0) {
+            console.log('Cant get flightInfo', err);
+          }
+          Promise.reject(err);
+        });
+    },
   };
 };
